+++ src/components/layout.tsx
@@ -8,9 +8,8 @@ import { ThemeProvider } from '@vtex/brand-ui'
 import styles from 'styles/documentation-page'
 import Header from 'components/header'
 import Footer from 'components/footer'
-import AnnouncementBar from './announcement-bar'
 
-import { DocumentationTitle, UpdatesTitle } from 'utils/typings/unionTypes'
+import { SectionId } from 'utils/typings/unionTypes'
 import Script from 'next/script'
 import {
   CookieBar,
@@ -18,9 +17,8 @@ import {
   Sidebar,
 } from '@vtexdocs/components'
 import {
-  documentationData,
+  menuDocumentationData,
   feedbackSectionData,
-  knownIssuesData,
   menuSupportData,
   updatesData,
 } from 'utils/constants'
@@ -32,7 +30,7 @@ interface Props {
   children: ReactElement
   hideSidebar?: boolean
   isPreview: boolean
-  sectionSelected?: DocumentationTitle | UpdatesTitle | ''
+  sectionSelected?: SectionId | ''
   parentsArray?: string[]
 }
 
@@ -66,12 +64,12 @@ export default function Layout({
     <ThemeProvider>
       <LibraryContextProvider
         sections={[
-          documentationData(intl),
-          knownIssuesData(intl),
+          menuDocumentationData(intl),
+          menuSupportData(intl),
           updatesData(intl),
         ]}
         hamburguerMenuSections={[
-          documentationData(intl),
+          menuDocumentationData(intl),
           menuSupportData(intl),
           updatesData(intl),
           feedbackSectionData(intl),
@@ -97,19 +95,6 @@ export default function Layout({
 					`}
           </Script>
         </div>
-        <AnnouncementBar
-          type="info"
-          label={intl.formatMessage({
-            id: 'announcement_bar.label',
-          })}
-          closable={false}
-          action={{
-            button: intl.formatMessage({
-              id: 'announcement_bar.button',
-            }),
-            href: 'https://help.vtex.com?utm_source=new-help-center-announcement-bar',
-          }}
-        />
         <CookieBar
           onAccept={() => {
             // eslint-disable-next-line @typescript-eslint/no-explicit-any