+++ src/pages/known-issues/[slug].tsx
@@ -1,16 +1,7 @@
 import Head from 'next/head'
 import { useEffect, useState, useContext, useRef } from 'react'
 import { GetStaticPaths, GetStaticProps, NextPage } from 'next'
-import { PHASE_PRODUCTION_BUILD } from 'next/constants'
-import { serialize } from 'next-mdx-remote/serialize'
 import { MDXRemoteSerializeResult } from 'next-mdx-remote'
-import remarkGFM from 'remark-gfm'
-import rehypeHighlight from 'rehype-highlight'
-import hljsCurl from 'highlightjs-curl'
-import remarkBlockquote from 'utils/remark_plugins/rehypeBlockquote'
-import { remarkCodeHike } from '@code-hike/mdx'
-import theme from 'styles/code-hike-theme'
-import remarkImages from 'utils/remark_plugins/plaiceholder'
 
 import { Box, Flex, Text } from '@vtex/brand-ui'
 
@@ -22,10 +13,8 @@ import { Item, TableOfContents } from '@vtexdocs/components'
 import Breadcrumb from 'components/breadcrumb'
 import TimeToRead from 'components/TimeToRead'
 
-import getHeadings from 'utils/getHeadings'
 import getNavigation from 'utils/getNavigation'
 // import getGithubFile from 'utils/getGithubFile'
-import escapeCurlyBraces from 'utils/escapeCurlyBraces'
 import replaceHTMLBlocks from 'utils/replaceHTMLBlocks'
 import { PreviewContext } from 'utils/contexts/preview'
 
@@ -33,28 +22,34 @@ import styles from 'styles/documentation-page'
 import { ContributorsType } from 'utils/getFileContributors'
 
 import { getLogger } from 'utils/logging/log-util'
-import { localeType } from 'utils/navigation-utils'
+import {
+  flattenJSON,
+  getArticleParentsArray,
+  getKeyByValue,
+} from 'utils/navigation-utils'
 import { MarkdownRenderer } from '@vtexdocs/components'
 // import { ParsedUrlQuery } from 'querystring'
 
-import { remarkReadingTime } from 'utils/remark_plugins/remarkReadingTime'
-import { getDocsPaths as getKnownIssuesPaths } from 'utils/getDocsPaths'
 import { getMessages } from 'utils/get-messages'
 import Tag from 'components/tag'
 import DateText from 'components/date-text'
 import CopyLinkButton from 'components/copy-link-button'
+import { serializeWithFallback } from 'utils/serializeWithFallback'
+import { extractStaticPropsParams } from 'utils/extractStaticPropsParams'
+import redirectToLocalizedUrl from 'utils/redirectToLocalizedUrl'
+import { fetchRawMarkdown } from 'utils/fetchRawMarkdown'
+import { fetchFileContributors } from 'utils/fetchFileContributors'
 
 // Initialize in getStaticProps
-let docsPathsGLOBAL: Record<string, { locale: string; path: string }[]> | null =
-  null
+const docsPathsGLOBAL: Record<
+  string,
+  { locale: string; path: string }[]
+> | null = null
 
 interface Props {
-  sectionSelected: string
   breadcrumbList: { slug: string; name: string; type: string }[]
-  content: string
   serialized: MDXRemoteSerializeResult
   contributors: ContributorsType[]
-  path: string
   headingList: Item[]
   branch: string
 }
@@ -90,7 +85,13 @@ const KnownIssuePage: NextPage<Props> = ({
     <>
       <Head>
         <title>{serialized.frontmatter?.title as string}</title>
-        <meta name="docsearch:doctype" content="Start here" />
+        <meta name="docsearch:doctype" content="known-issues" />
+        {serialized.frontmatter?.title && (
+          <meta
+            name="docsearch:doctitle"
+            content={serialized.frontmatter.title as string}
+          />
+        )}
         {serialized.frontmatter?.hidden && (
           <meta name="robots" content="noindex" />
         )}
@@ -187,150 +188,85 @@ export const getStaticProps: GetStaticProps = async ({
   preview,
   previewData,
 }) => {
-  const previewBranch =
-    preview && JSON.parse(JSON.stringify(previewData)).hasOwnProperty('branch')
-      ? JSON.parse(JSON.stringify(previewData)).branch
-      : 'main'
-  const branch = preview ? previewBranch : 'main'
-  const slug = params?.slug as string
-  const currentLocale: localeType = locale
-    ? (locale as localeType)
-    : ('en' as localeType)
-  if (!docsPathsGLOBAL) {
-    docsPathsGLOBAL = await getKnownIssuesPaths('known-issues')
-  }
-  const docsPaths =
-    preview || process.env.NEXT_PHASE !== PHASE_PRODUCTION_BUILD
-      ? await getKnownIssuesPaths('known-issues', branch)
-      : docsPathsGLOBAL
-
-  const logger = getLogger('Start here')
-
-  const slugPaths = docsPaths[slug]
-
-  if (!Array.isArray(slugPaths)) {
-    logger.warn(
-      `Slug not found in docsPaths: slug: ${slug}, locale: ${currentLocale}, branch: ${branch}`
-    )
-    return {
-      notFound: true,
-    }
-  }
-
-  const path = slugPaths.find((e) => e.locale === currentLocale)?.path
-
-  if (!path) {
-    logger.warn(
-      `File exists in the repo but not in navigation: slug: ${slug}, locale: ${currentLocale}, branch: ${branch}`
-    )
-    return {
-      notFound: true,
-    }
-  }
-
-  if (!path) {
-    logger.warn(
-      `File exists in the repo but not in navigation: slug: ${slug}, locale: ${currentLocale}, branch: ${branch}`
-    )
-    return {
-      notFound: true,
-    }
-  }
-
-  let documentationContent =
-    (await fetch(
-      `https://raw.githubusercontent.com/vtexdocs/help-center-content/${branch}/${path}`
-    )
-      .then((res) => res.text())
-      .catch((err) => console.log(err))) || ''
-
-  // Serialize content and parse frontmatter
-  const serialized = await serialize(documentationContent, {
-    parseFrontmatter: true,
-  })
-
-  // Allow PUBLISHED and CHANGED status documents to be visible
-  const allowedStatuses = ['PUBLISHED', 'CHANGED']
-  const hasAllowedStatus = allowedStatuses.includes(
-    serialized?.frontmatter?.status as string
-  )
-  if (!hasAllowedStatus) {
-    return {
-      notFound: true,
-    }
-  }
-
-  const contributors =
-    (await fetch(
-      `https://github.com/vtexdocs/help-center-content/file-contributors/${branch}/${path}`,
-      {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-        },
-      }
-    )
-      .then((res) => res.json())
-      .then(({ users }) => {
-        const result: ContributorsType[] = []
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        for (let i = 0; i < users.length; i++) {
-          const user = users[i]
-          if (user.id === '41898282') continue
-          result.push({
-            name: user.login,
-            login: user.login,
-            avatar: user.primaryAvatarUrl,
-            userPage: `https://github.com${user.profileLink}`,
-          })
-        }
-
-        return result
-      })
-      .catch((err) => console.log(err))) || []
-
-  let format: 'md' | 'mdx' = 'mdx'
-  try {
-    if (path.endsWith('.md')) {
-      documentationContent = escapeCurlyBraces(documentationContent)
-      documentationContent = replaceHTMLBlocks(documentationContent)
-    }
-  } catch (error) {
-    logger.error(`${error}`)
-    format = 'md'
-  }
-
-  try {
-    const headingList: Item[] = []
-    let serialized = await serialize(documentationContent, {
-      parseFrontmatter: true,
-      mdxOptions: {
-        remarkPlugins: [
-          [remarkCodeHike, theme],
-          remarkGFM,
-          remarkImages,
-          [getHeadings, { headingList }],
-          remarkBlockquote,
-          remarkReadingTime,
-        ],
-        useDynamicImport: true,
-        rehypePlugins: [
-          [rehypeHighlight, { languages: { hljsCurl }, ignoreMissing: true }],
-        ],
-        format,
-      },
+  const logger = getLogger('Known Issues')
+  const { sectionSelected, branch, slug, currentLocale, docsPaths, docExists } =
+    await extractStaticPropsParams({
+      sectionSelected: 'known-issues',
+      params,
+      locale,
+      preview,
+      previewData,
+      docsPathsGLOBAL,
     })
 
+  if (!docExists) {
+    logger.warn(
+      `Markdown file not found for slug: ${slug}, locale: ${currentLocale}, branch: ${branch}`
+    )
+    return { notFound: true }
+  }
+  try {
+    const path = docsPaths[slug]?.find((e) => e.locale === currentLocale)?.path
     const sidebarfallback = await getNavigation()
-    serialized = JSON.parse(JSON.stringify(serialized))
+    const filteredSidebar = sidebarfallback.find(
+      (item: { documentation: string }) =>
+        item.documentation === sectionSelected
+    )
+    const flattenedSidebar = flattenJSON(filteredSidebar)
+    const keyPath = getKeyByValue(flattenedSidebar, slug) as string
+
+    if (!path) {
+      logger.warn(
+        `Markdown file (slug: ${slug}, locale: ${currentLocale}, branch: ${branch}) exists for another locale. Redirecting to localized version.`
+      )
+      return redirectToLocalizedUrl(
+        keyPath,
+        currentLocale,
+        flattenedSidebar,
+        'known-issues'
+      )
+    }
+
+    const rawContent = await fetchRawMarkdown(branch, path)
+    const documentationContent = replaceHTMLBlocks(rawContent)
+
+    let parentsArray: string[] = []
+
+    if (keyPath) {
+      parentsArray = getArticleParentsArray(
+        keyPath,
+        flattenedSidebar,
+        currentLocale,
+        slug
+      )
+    }
+
+    const headingList: Item[] = []
+    const serialized = await serializeWithFallback({
+      content: documentationContent,
+      headingList,
+      logger,
+      path,
+    })
+    if (!serialized) {
+      logger.error(`Serialization failed for ${path}`)
+      return { notFound: true }
+    }
+
+    const status = serialized.frontmatter?.status as string
+    if (!['PUBLISHED', 'CHANGED'].includes(status)) {
+      logger.warn(
+        `Document status is not allowed for ${path}. Status: ${status}.'
+        )}`
+      )
+      return { notFound: true }
+    }
+
+    const contributors = await fetchFileContributors(branch, path)
 
     logger.info(`Processing ${slug}`)
-    const seeAlsoData: {
-      url: string
-      title: string
-      category: string
-    }[] = []
+
+    const seeAlsoData: { url: string; title: string; category: string }[] = []
 
     const breadcrumbList: { slug: string; name: string; type: string }[] = [
       {
@@ -347,10 +283,12 @@ export const getStaticProps: GetStaticProps = async ({
 
     return {
       props: {
+        sectionSelected,
         slug,
-        serialized,
+        serialized: JSON.parse(JSON.stringify(serialized)),
         sidebarfallback,
         headingList,
+        parentsArray,
         contributors,
         path,
         seeAlsoData,
@@ -360,10 +298,8 @@ export const getStaticProps: GetStaticProps = async ({
       revalidate: 600,
     }
   } catch (error) {
-    logger.error(`Error while processing ${path}\n${error}`)
-    return {
-      notFound: true,
-    }
+    logger.error(`Error while processing ${params?.slug}:\n${error}`)
+    return { notFound: true }
   }
 }