+++ src/utils/logging/log-util.ts
@@ -59,3 +59,9 @@ export function getLogger(name: string) {
     },
   }
 }
+
+export type LoggerType = {
+  error: (msg: string) => void
+  info: (msg: string) => void
+  warn: (msg: string) => void
+}